Adding effects finished. They are saved in the database with many to many relationships with heroes and leaders. Bugs are accepted Working fine!

+++ src/main/java/com/saidashevar/ptgame/model/Turn.java
@@ -46,7 +46,7 @@ public class Turn {
 	public byte getActionsLeft() {
 		return actionsLeft;
 	}
-	public void setActionsLeft(byte actionsLeft) {
-		this.actionsLeft = actionsLeft;
+	public void setActionsLeft(int actionsLeft) {
+		this.actionsLeft = (byte)actionsLeft;
 	}
 }
